Added response compression to improve performance
Added the `Microsoft.AspNetCore.ResponseCompression` and `System.IO.Compression` namespaces to the `Program.cs` file. A new service for response compression has been added to the application's service collection using the `AddResponseCompression` method. The `EnableForHttps` property is set to `true`, enabling the response compression middleware for HTTPS requests. Configured `BrotliCompressionProviderOptions` with a compression level set to `Fastest` to reduce the size of transmitted data.